requirements.txt
requests==2.31.0
argparse==1.4.0
pymongo==4.5.0
python-dotenv==1.0.0
authlib==1.2.1
click==8.1.7
rich==13.6.0
tabulate==0.9.0
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0 